CHG Pages Refactoring
