+++ web/src/Pages/Public/Courses/Courses.styles.jsx
@@ -1,0 +1,7 @@
+import {makeStyles} from '@material-ui/core';
+
+export const useStyles = makeStyles((theme) => ({
+  inline: {
+    display: 'inline',
+  },
+}));
