+++ web/src/Pages/Public/Repos.jsx
@@ -1,10 +1,6 @@
 import React, {useState} from 'react';
-import useSWR from 'swr';
-import {Redirect} from 'react-router-dom';
 
 import Grid from '@material-ui/core/Grid';
-import CircularProgress from '@material-ui/core/CircularProgress';
-
 
 import ReposTable from '../../Components/Public/Repos/ReposTable';
 import StandardLayout from '../../Components/Layouts/StandardLayout';
@@ -15,7 +11,7 @@ import {useSnackbar} from 'notistack';
 import standardErrorHandler from '../../Utils/standardErrorHandler';
 
 
-export default function Repos() {
+const Repos = () => {
   const [rows, setRows] = useState([]);
   const [reset, setReset] = useState(0);
   const {enqueueSnackbar} = useSnackbar();
@@ -42,4 +38,7 @@ export default function Repos() {
       )}
     </AuthContext.Consumer>
   );
-}
+};
+
+export default Repos;
+
