+++ web/src/Pages/Public/Visuals/Visuals.jsx
@@ -1,33 +1,17 @@
 import React from 'react';
 
-import makeStyles from '@material-ui/core/styles/makeStyles';
 import Grid from '@material-ui/core/Grid';
 import Card from '@material-ui/core/Card';
 import CardMedia from '@material-ui/core/CardMedia';
 import CardHeader from '@material-ui/core/CardHeader';
 import Avatar from '@material-ui/core/Avatar';
 
-import StandardLayout from '../../Components/Layouts/StandardLayout';
+import {useStyles} from './Visuals.styles';
+import StandardLayout from '../../../Components/Layouts/StandardLayout';
 import Button from '@material-ui/core/Button';
 import CloudDownload from '@material-ui/icons/CloudDownload';
 
-
-const useStyles = makeStyles((theme) => ({
-  usage: {
-    height: 0,
-    paddingTop: '83.33%', // 16:9
-  },
-  title: {
-    padding: theme.spacing(1, 2),
-    fontSize: 16,
-  },
-  button: {
-    margin: theme.spacing(2, 1),
-  },
-}));
-
-
-export default function Visuals() {
+const Visuals = () => {
   const classes = useStyles();
 
   return (
@@ -63,4 +47,6 @@ export default function Visuals() {
       </Grid>
     </StandardLayout>
   );
-}
+};
+
+export default Visuals;
