+++ web/src/navconfig.jsx
@@ -18,15 +18,15 @@ import DashboardIcon from '@material-ui/icons/Dashboard';
 
 import Dashboard from './Pages/Public/Dashboard/Dashboard';
 import About from './Pages/Public/About';
-import Courses from './Pages/Public/Courses';
+import Courses from './Pages/Public/Courses/Courses';
 import Assignments from './Pages/Public/Assignments';
 import Profile from './Pages/Public/Profile';
 import Repos from './Pages/Public/Repos';
 import Submissions from './Pages/Public/Submissions';
-import Submission from './Pages/Public/Submission';
+import Submission from './Pages/Public/Submission/Submission';
 import Blog from './Pages/Public/Blog';
-import Visuals from './Pages/Public/Visuals';
-import Lectures from './Pages/Public/Lectures';
+import Visuals from './Pages/Public/Visuals/Visuals';
+import Lectures from './Pages/Public/Lectures/Lectures';
 
 import AdminUsers from './Pages/Admin/Users';
 import AdminUser from './Pages/Admin/User';
